+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## VNext
+- Added assert for connection string
+
 ## 0.1.7-cere Jul 20, 2021
 - Added support for External Dev cluster