+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # CHANGELOG
 
 ## VNext 
-- Updating URL validation
+- Updated URL validation
 
 ## 0.1.7-cere Jul 20, 2021
 - Added support for External Dev cluster