+++ CHANGELOG.md
@@ -2,8 +2,8 @@
 
 
 ## vNext 
-- Updated URL validation
 - Added assert for connection string
+- Updated URL validation
 
 ## 0.1.7-cere Jul 20, 2021
 - Added support for External Dev cluster