+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # CHANGELOG
 
-## VNext
+
+## vNext 
 - Added assert for connection string
+- Updated URL validation
 
 ## 0.1.7-cere Jul 20, 2021
 - Added support for External Dev cluster